refactor: extract GithubClientFinder concern and fix seeds loading
- Extract find_github_client to shared Concern module
- Remove duplicate code from 3 tool services
- Fix seeds.rb to run on rails db:seed (remove  check)
