+++ engines/collavre_github/app/services/collavre_github/tools/github_pr_details_service.rb
@@ -6,6 +6,7 @@ module CollavreGithub
     class GithubPrDetailsService
       extend T::Sig
       extend ToolMeta
+      include Concerns::GithubClientFinder
 
       tool_name "github_pr_details"
       tool_description <<~DESC.strip
@@ -44,25 +45,6 @@ module CollavreGithub
         }
       rescue StandardError => e
         { error: "Failed to fetch PR details: #{e.message}" }
-      end
-
-      private
-
-      def find_github_client(creative_id, repo)
-        creative = Collavre::Creative.find_by(id: creative_id)
-        return nil unless creative
-
-        # Find repository link for this creative (or its origin)
-        origin = creative.effective_origin
-        link = CollavreGithub::RepositoryLink
-          .joins(:creative)
-          .where(repository_full_name: repo)
-          .where(creative: origin.self_and_descendants)
-          .first
-
-        return nil unless link&.github_account
-
-        CollavreGithub::Client.new(link.github_account)
       end
     end
   end
