+++ engines/collavre_github/app/services/collavre_github/tools/github_pr_diff_service.rb
@@ -6,6 +6,7 @@ module CollavreGithub
     class GithubPrDiffService
       extend T::Sig
       extend ToolMeta
+      include Concerns::GithubClientFinder
 
       DIFF_MAX_LENGTH = 10_000
 
@@ -48,24 +49,6 @@ module CollavreGithub
         }
       rescue StandardError => e
         { error: "Failed to fetch PR diff: #{e.message}" }
-      end
-
-      private
-
-      def find_github_client(creative_id, repo)
-        creative = Collavre::Creative.find_by(id: creative_id)
-        return nil unless creative
-
-        origin = creative.effective_origin
-        link = CollavreGithub::RepositoryLink
-          .joins(:creative)
-          .where(repository_full_name: repo)
-          .where(creative: origin.self_and_descendants)
-          .first
-
-        return nil unless link&.github_account
-
-        CollavreGithub::Client.new(link.github_account)
       end
     end
   end
